Move mocks and fixtures into test utils
